fixed powerlevel9k

+++ install.zsh
@@ -4,6 +4,11 @@ cp .zshrc $HOME
 if [ ! -d "${HOME}/.oh-my-zsh" ]; then
   sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 fi
-rm -r $HOME/.zsh
+if [ ! -d "${HOME}/.oh-my-zsh/custom/themes/powerlevel9k" ]; then
+  git clone https://github.com/bhilburn/powerlevel9k.git ~/.oh-my-zsh/custom/themes/powerlevel9k
+fi
+if [ -d "${HOME}/.zsh" ]; then
+  rm -r $HOME/.zsh
+fi
 cp -r .zsh $HOME
 source ~/.zshrc
